+++ src/core/traits/insertion_algorithm.rs
@@ -909,9 +909,9 @@ where
                 let (cell_key, facet_index) = cells[0];
                 if let Some(cell) = tds.cells().get(cell_key)
                     && let Ok(facets) = cell.facets()
-                    && facet_index < facets.len()
+                    && (facet_index as usize) < facets.len()
                 {
-                    let facet = &facets[facet_index];
+                    let facet = &facets[facet_index as usize];
 
                     // Try to create a cell from this facet and the vertex
                     if Self::create_cell_from_facet_and_vertex(tds, facet, vertex) {
@@ -932,9 +932,9 @@ where
             for &(cell_key, facet_index) in cells {
                 if let Some(cell) = tds.cells().get(cell_key)
                     && let Ok(facets) = cell.facets()
-                    && facet_index < facets.len()
+                    && (facet_index as usize) < facets.len()
                 {
-                    let facet = &facets[facet_index];
+                    let facet = &facets[facet_index as usize];
 
                     // Try to create a cell from this facet and the vertex
                     if Self::create_cell_from_facet_and_vertex(tds, facet, vertex) {
@@ -1190,8 +1190,8 @@ where
             let (cell_key, facet_index) = cells[0];
             if let Some(cell) = tds.cells().get(cell_key) {
                 if let Ok(facets) = cell.facets() {
-                    if facet_index < facets.len() {
-                        let facet = &facets[facet_index];
+                    if (facet_index as usize) < facets.len() {
+                        let facet = &facets[facet_index as usize];
 
                         // Test visibility using proper orientation predicates
                         if Self::is_facet_visible_from_vertex_impl(tds, facet, vertex, cell_key) {
